refactor : Move RecipeImage in ui utils
